fix(mg-button): variant not updating when not explicitly set and displayed conditionally
#503 CORE-3435

+++ packages/mg-components/src/components/atoms/mg-button/mg-button.tsx
@@ -32,17 +32,20 @@ export class MgButton {
   /**
    * Define button variant
    */
-  @Prop() variant: VariantType = variants[0]; // Primary
+  @Prop({ mutable: true }) variant: VariantType = variants[0]; // Primary
   @Watch('variant')
   validateVariant(newValue: VariantType, oldValue?: VariantType): void {
-    if (!variants.includes(newValue)) {
+    // validate new value
+    if (typeof newValue === 'string' && newValue.length === 0) {
+      // Reactive framework like VUE render with `""` when prop is update to undefined
+      this.variant = variants[0];
+    } else if (!variants.includes(newValue)) {
       throw new Error(`<mg-button> prop "variant" must be one of: ${variants.join(', ')}. Passed value: ${toString(newValue)}.`);
-    } else {
-      if (oldValue !== undefined) {
-        this.classCollection.delete(`mg-c-button--${oldValue}`);
-      }
-      this.classCollection.add(`mg-c-button--${newValue}`);
     }
+
+    // manage class updates
+    if (variants.includes(oldValue)) this.classCollection.delete(`mg-c-button--${oldValue}`);
+    if (variants.includes(newValue)) this.classCollection.add(`mg-c-button--${newValue}`);
   }
 
   /**
